# Conflicts:
#	common/src/main/java/com/wynntils/features/map/MainMapFeature.java
#	common/src/main/java/com/wynntils/screens/maps/AbstractMapScreen.java
#	common/src/main/java/com/wynntils/services/map/PoiService.java

+++ common/src/main/java/com/wynntils/models/damage/type/DamageDealtEvent.java
@@ -1,12 +1,12 @@
 /*
- * Copyright © Wynntils 2023.
+ * Copyright © Wynntils 2023-2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.models.damage.type;
 
 import com.wynntils.models.stats.type.DamageType;
 import java.util.Map;
-import net.minecraftforge.eventbus.api.Event;
+import net.neoforged.bus.api.Event;
 
 /**
  * This event is sent out whenever Wynntils register that a mob has received damage.
@@ -14,13 +14,13 @@ import net.minecraftforge.eventbus.api.Event;
  * spell attack from the current player.
  */
 public final class DamageDealtEvent extends Event {
-    private final Map<DamageType, Integer> damages;
+    private final Map<DamageType, Long> damages;
 
-    public DamageDealtEvent(Map<DamageType, Integer> damages) {
+    public DamageDealtEvent(Map<DamageType, Long> damages) {
         this.damages = damages;
     }
 
-    public Map<DamageType, Integer> getDamages() {
+    public Map<DamageType, Long> getDamages() {
         return damages;
     }
 }
